requirements.txt
# PyTorch - Machine Learning Framework
torch>=2.0.0

# OpenAI Whisper - Speech Recognition Model
openai-whisper>=20230918

# (Optional) Rich for improved console output
rich>=13.5.0